+++ pytest_tars/_srvdef.py
@@ -21,19 +21,19 @@ __all__ = ['ServerDefinition']
 
 class ServerDefinition(object):
     """
-    Encapsulation of a server definition from a server definition file.
+    Represents a single server definition from a server definition file.
 
     Example for a server definition in a server definition file:
 
     .. code-block:: yaml
 
-        myserver1:                        # nickname of the server
+        myserver1:                              # nickname of the server
           description: "my dev system 1"
           contact_name: "John Doe"
           access_via: "VPN to dev network"
-          details:                        # user-defined part
+          user_defined:                         # user-defined part
             host: "10.11.12.13"
-            userid: myuserid
+            username: myusername
             password: mypassword
             stuff:
               - morestuff1
@@ -44,7 +44,7 @@ class ServerDefinition(object):
         self._description = server_dict['description']
         self._contact_name = server_dict.get('contact_name', None)
         self._access_via = server_dict.get('access_via', None)
-        self._details = server_dict['details']
+        self._user_defined = server_dict['user_defined']
 
     def __repr__(self):
         return "ServerDefinition(" \
@@ -52,7 +52,7 @@ class ServerDefinition(object):
             "description={s._description!r}, " \
             "contact_name={s._contact_name!r}, " \
             "access_via={s._access_via!r}, " \
-            "details={s._details!r})". \
+            "user_defined={s._user_defined!r})". \
             format(s=self)
 
     @property
@@ -96,12 +96,12 @@ class ServerDefinition(object):
         return self._access_via
 
     @property
-    def details(self):
+    def user_defined(self):
         """
         dict: Details of the server, such as IP address.
 
-        This is the value of the ``details`` property of the server object
+        This is the value of the ``user_defined`` property of the server object
         in the server definition file. This value can have an arbitrary
         user-defined structure.
         """
-        return self._details
+        return self._user_defined